requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
qrcode[pil]==7.4.2
pillow==10.1.0
python-dotenv==1.0.0
alembic==1.12.1
psycopg2-binary==2.9.9 
Flask==3.0.3
Werkzeug==3.0.3
cloudinary==1.41.0